+++ src/Core/SymbolUtility.cs
@@ -575,8 +575,8 @@ namespace Roslynator
                     return methodSymbol;
 
                 typeSymbol = typeSymbol.BaseType;
-
-            } while (typeSymbol != null
+            }
+            while (typeSymbol != null
                 && typeSymbol.SpecialType != SpecialType.System_Object);
 
             return null;
@@ -620,12 +620,6 @@ namespace Roslynator
                     case "ValueTask`1":
                         return true;
                 }
-            }
-
-            if (typeKind == TypeKind.Interface
-                && originalDefinition.HasMetadataName(MetadataNames.System_Collections_Generic_IAsyncEnumerable_T))
-            {
-                return true;
             }
 
             if (typeKind == TypeKind.Class